// src/types/auth-context.interface.ts
import { User } from "./user.interface";

export interface AuthContextData {
  user: User | null;
  accessToken: string | null;
  refreshToken: string | null;
  login: (email: string, password: string) => Promise<void>;
  logout: () => void;
  register: (email: string, password: string, name: string) => Promise<void>;
  refreshAccessToken: () => Promise<void>;
}